crawler/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
playwright>=1.40.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
